Fixed the underline issue
When using Sogou input method, there will be underline when user typing, but actually user doesn't really click the Underline icon to make it underline, so this is to fix the issue. BTW, clearComposingText() will cause side effect which is when user typing a character, there willl be two come out.